Add ability to validate parameters in components
Adds a new `Access` object that tracks the access requirements of a
component by allowing each of its parameters to register itself with the
component's Access object, effectively self-validating.

Many updates to the workspace revolve around adding this support.

Mainly: Removes the `Service` ComponentParam as it was just an example.
Will add a replacement down the line.