Add `Protocol` ComponentParam and supporting code
Adds a new ComponentParam, `Protocol`, which is used to request a
protocol from the underlying storage. If the protocol is not available,
the component will not be run.

Implements the `ComponentParam` trait to the `Storage` struct to allow
for the creation of data (For now Protocols) inside of a component, and
not just as pre-registration before running all components. This will
not be the final implementation as it requires exclusive access to
storage, and thus cannot be parallelized.